增加添加字典和字典项的相关规则，使得后端能够接收字典编码的字段并返回字典具体数据。

+++ src/main/java/io/renren/modules/business/entity/OrderPartRelEntity.java
@@ -40,7 +40,7 @@ public class OrderPartRelEntity implements Serializable {
 	/**
 	 * 使用数量
 	 */
-	private Integer usedQuantity;
+	private BigDecimal usedQuantity;
 	/**
 	 * 配件总价
 	 */
